DRF Tutorial Series - Split views.py in separate modules
